backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain==0.1.0
langchain-community==0.0.10
chromadb
pypdf2
pypdf
python-dotenv
pydantic==1.10.13
sentence-transformers
faiss-cpu
numpy
pandas
transformers
torch
torchvision
accelerate
huggingface-hub